+++ src/main/java/io/github/vantiv/sdk/Versions.java
@@ -2,6 +2,6 @@ package io.github.vantiv.sdk;
 
 public class Versions {
 
-    public static final String XML_VERSION = "12.50";
-    public static final String SDK_VERSION = "Java;12.50.0";
+    public static final String XML_VERSION = "12.51";
+    public static final String SDK_VERSION = "Java;12.51.0";
 }